Move the searches db code into a new file
